requirements.txt
streamlit
ultralytics
pillow
opencv-python-headless
numpy
pyyaml
